added stick

+++ qq_party_shared/src/lib.rs
@@ -35,7 +35,7 @@ pub struct ChaseTargetId(pub u32, pub u8);//ball, npc, speed
 #[derive(Component,Serialize, Deserialize, Default, Clone, Copy, Debug, PartialEq, Hash, Eq)]
 pub struct ChaseTargetId2(pub u32, pub Option<Entity>,pub u8);//ball, npc, speed
 #[derive(Component,Serialize, Deserialize, Default, Clone, Debug)]
-pub struct FireId(pub u32,pub u32,pub Option<Vec2>); //owner, sprite_enum, StartPosition
+pub struct FireId(pub u32,pub i16,pub Option<Vec2>); //owner, sprite_enum, StartPosition
 #[derive(Component,Serialize, Deserialize, Default, Clone, Debug, PartialEq, Hash, Eq)]
 pub struct Hit;
 #[derive(Component,Serialize, Deserialize, Default, Clone, Copy, Debug, PartialEq, Hash, Eq)]
